Refactor authentication and dashboard components: removed unused AuthProvider and AdminGuard, simplified LoginPage with new components, updated DashboardPage to display user names correctly, and enhanced RecentSales with accordion for better data presentation.
