+++ src/app/(admin)/layout.tsx
@@ -1,5 +1,4 @@
 import DashboardLayout from "@/components/layout/dashboard";
-import { AuthProvider } from "@/providers/auth-provider";
 
 export default function AdminLayout({
   children,
@@ -7,10 +6,8 @@ export default function AdminLayout({
   children: React.ReactNode;
 }) {
   return (
-    <AuthProvider>
-      <DashboardLayout>
-        <div className="p-4">{children}</div>
-      </DashboardLayout>
-    </AuthProvider>
+    <DashboardLayout>
+      <div className="p-4">{children}</div>
+    </DashboardLayout>
   );
 }
