+++ src/app/(auth)/login/page.tsx
@@ -1,41 +1,22 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
 import {
   Card,
   CardHeader,
   CardTitle,
   CardDescription,
-  CardContent,
-  CardFooter,
 } from "@/components/ui/card";
-import { User, Lock, CheckCircle2 } from "lucide-react";
-import Link from "next/link";
 import { toast } from "sonner";
-import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
-import { motion } from "framer-motion";
+import { LoginForm } from "@/components/login/LoginForm";
+import { LoggedInRedirect } from "@/components/login/LoggedInRedirect";
+import { LoginBackground } from "@/components/login/LoginBackground";
+import { loginSchema, LoginFormData } from "@/components/login/types";
 import { LoginCredentials } from "@/types/auth";
 import { useAuth } from "@/contexts/AuthContext";
-
-const loginSchema = z.object({
-  phone_number: z.string().min(1, "Phone number is required"),
-  password: z.string(),
-});
-
-type LoginFormData = z.infer<typeof loginSchema>;
 
 export default function LoginPage() {
   const { login, user } = useAuth();
@@ -56,7 +37,6 @@ export default function LoginPage() {
       const timeout = setTimeout(() => {
         router.push("/dashboard");
       }, 2000);
-
       return () => clearTimeout(timeout);
     }
   }, [user, router]);
@@ -64,7 +44,6 @@ export default function LoginPage() {
   const onSubmit = async (data: LoginFormData) => {
     setIsLoading(true);
     setErrorMessage(null);
-
     try {
       const loginData: LoginCredentials = {
         phone_number: data.phone_number,
@@ -74,6 +53,7 @@ export default function LoginPage() {
       toast.success("Login successful!");
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (error: any) {
+      console.log(error);
       const errorMessage =
         error.response?.data?.error || "An unexpected error occurred.";
       setErrorMessage(errorMessage);
@@ -84,84 +64,12 @@ export default function LoginPage() {
   };
 
   if (user) {
-    return (
-      <div className="min-h-screen flex items-center justify-center relative overflow-hidden">
-        <div className="absolute inset-0 bg-gradient-to-br from-blue-50 via-indigo-50 to-purple-50">
-          <div className="absolute inset-0 bg-grid-slate-200 [mask-image:linear-gradient(0deg,white,rgba(255,255,255,0.6))]" />
-          <div className="absolute inset-0 bg-gradient-to-br from-transparent via-purple-100/30 to-blue-100/30 animate-gradient" />
-        </div>
-
-        <motion.div
-          initial={{ scale: 0.8, opacity: 0 }}
-          animate={{ scale: 1, opacity: 1 }}
-          transition={{ duration: 0.5, ease: "easeOut" }}
-          className="relative z-10"
-        >
-          <Card className="w-[350px] shadow-xl bg-white/80 backdrop-blur-sm">
-            <CardContent className="pt-6">
-              <div className="flex flex-col items-center space-y-4">
-                <motion.div
-                  initial={{ scale: 0 }}
-                  animate={{ scale: 1 }}
-                  transition={{ delay: 0.2, type: "spring", stiffness: 200 }}
-                >
-                  <div className="rounded-full bg-green-100 p-3">
-                    <CheckCircle2 className="w-8 h-8 text-green-600" />
-                  </div>
-                </motion.div>
-
-                <motion.div
-                  initial={{ y: 20, opacity: 0 }}
-                  animate={{ y: 0, opacity: 1 }}
-                  transition={{ delay: 0.3 }}
-                  className="text-center space-y-2"
-                >
-                  <h2 className="text-2xl font-semibold text-gray-800">
-                    Welcome Back!
-                  </h2>
-                  <p className="text-gray-600">You&re already logged in</p>
-                </motion.div>
-
-                <motion.div
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  transition={{ delay: 0.5 }}
-                  className="w-full"
-                >
-                  <div className="relative pt-4">
-                    <div className="w-full bg-gray-200 rounded-full h-1.5">
-                      <motion.div
-                        initial={{ width: "0%" }}
-                        animate={{ width: "100%" }}
-                        transition={{ duration: 2, ease: "easeInOut" }}
-                        className="bg-green-500 h-1.5 rounded-full"
-                      />
-                    </div>
-                  </div>
-                  <p className="text-sm text-gray-500 text-center mt-2">
-                    Redirecting to dashboard...
-                  </p>
-                </motion.div>
-              </div>
-            </CardContent>
-          </Card>
-        </motion.div>
-      </div>
-    );
+    return <LoggedInRedirect />;
   }
 
   return (
     <div className="min-h-screen flex items-center justify-center relative overflow-hidden">
-      <div className="absolute inset-0 bg-gradient-to-br from-blue-50 via-indigo-50 to-purple-50">
-        <div className="absolute inset-0 bg-grid-slate-200 [mask-image:linear-gradient(0deg,white,rgba(255,255,255,0.6))]" />
-        <div className="absolute inset-0 bg-gradient-to-br from-transparent via-purple-100/30 to-blue-100/30 animate-gradient" />
-      </div>
-      <div className="absolute inset-0 overflow-hidden pointer-events-none">
-        <div className="absolute -top-4 -left-4 w-72 h-72 bg-purple-200 rounded-full mix-blend-multiply filter blur-xl opacity-70 animate-blob" />
-        <div className="absolute -bottom-8 -right-4 w-72 h-72 bg-blue-200 rounded-full mix-blend-multiply filter blur-xl opacity-70 animate-blob animation-delay-2000" />
-        <div className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 w-72 h-72 bg-indigo-200 rounded-full mix-blend-multiply filter blur-xl opacity-70 animate-blob animation-delay-4000" />
-      </div>
-
+      <LoginBackground />
       <Card className="w-full max-w-md mx-4 shadow-xl bg-white/80 backdrop-blur-sm relative z-10">
         <CardHeader className="space-y-4 text-center pb-8">
           <CardTitle className="text-3xl font-bold bg-gradient-to-r from-blue-800 to-purple-600 bg-clip-text text-transparent">
@@ -171,90 +79,12 @@ export default function LoginPage() {
             Sign in to your Yachu Sales App account
           </CardDescription>
         </CardHeader>
-
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-            <CardContent className="space-y-6">
-              <FormField
-                control={form.control}
-                name="phone_number"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel className="text-gray-700 font-medium">
-                      Phone Number
-                    </FormLabel>
-                    <FormControl>
-                      <div className="relative group">
-                        <User className="absolute left-3 top-3 h-4 w-4 text-gray-400 group-hover:text-purple-500 transition-colors" />
-                        <Input
-                          className="pl-10 h-12 border-gray-200 focus:border-purple-400 focus:ring-purple-400 transition-all"
-                          placeholder="Enter your phone number"
-                          {...field}
-                        />
-                      </div>
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel className="text-gray-700 font-medium">
-                      Password
-                    </FormLabel>
-                    <FormControl>
-                      <div className="relative group">
-                        <Lock className="absolute left-3 top-3 h-4 w-4 text-gray-400 group-hover:text-purple-500 transition-colors" />
-                        <Input
-                          type="password"
-                          className="pl-10 h-12 border-gray-200 focus:border-purple-400 focus:ring-purple-400 transition-all"
-                          placeholder="Enter your password"
-                          {...field}
-                        />
-                      </div>
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <div className="flex items-center justify-between text-sm">
-                <Link
-                  href="/forgot-password"
-                  className="text-blue-800 hover:text-purple-600 hover:underline font-medium transition-colors"
-                >
-                  Forgot password?
-                </Link>
-              </div>
-            </CardContent>
-
-            <CardFooter className="flex flex-col space-y-4">
-              <Button
-                type="submit"
-                className="w-full h-12 bg-gradient-to-r from-blue-800 to-purple-600 hover:from-blue-800 hover:to-purple-700 text-white text-lg font-medium transition-all duration-300 transform hover:scale-[1.02]"
-                disabled={isLoading}
-              >
-                {isLoading ? (
-                  <div className="flex items-center space-x-2">
-                    <div className="w-5 h-5 border-t-2 border-b-2 border-white rounded-full animate-spin" />
-                    <span>Signing in...</span>
-                  </div>
-                ) : (
-                  "Sign in"
-                )}
-              </Button>
-              {errorMessage && (
-                <div className="text-red-600 text-sm text-center mt-2">
-                  {errorMessage}
-                </div>
-              )}
-            </CardFooter>
-          </form>
-        </Form>
+        <LoginForm
+          isLoading={isLoading}
+          errorMessage={errorMessage}
+          onSubmit={onSubmit}
+          form={form}
+        />
       </Card>
     </div>
   );
