--- src/components/auth/admin-guard.tsx
@@ -1,50 +1,0 @@
-"use client";
-
-import { useAuth } from "@/providers/auth-provider";
-import { useEffect, useState } from "react";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import LoginForm from "@/components/auth/login-form";
-
-export default function AdminGuard({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const { isAuthenticated, loading } = useAuth();
-  const [showLoginDialog, setShowLoginDialog] = useState(false);
-
-  useEffect(() => {
-    if (!loading && !isAuthenticated) {
-      setShowLoginDialog(true);
-    } else {
-      setShowLoginDialog(false);
-    }
-  }, [isAuthenticated, loading]);
-
-  if (loading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center">
-        <div className="animate-spin rounded-full h-32 w-32 border-b-2 border-gray-900"></div>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <Dialog open={showLoginDialog} onOpenChange={setShowLoginDialog}>
-        <DialogContent className="sm:max-w-[425px]">
-          <DialogHeader>
-            <DialogTitle>Login Required</DialogTitle>
-          </DialogHeader>
-          <LoginForm />
-        </DialogContent>
-      </Dialog>
-      {isAuthenticated ? children : null}
-    </>
-  );
-}
