--- src/components/auth/login-form.tsx
@@ -1,56 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { useAuth } from "@/providers/auth-provider";
-
-export default function LoginForm() {
-  const [phoneNumber, setPhoneNumber] = useState("");
-  const [password, setPassword] = useState("");
-  const { login } = useAuth();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    try {
-      await login(phoneNumber, password);
-    } catch (error) {
-      console.error("Login failed:", error);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="space-y-4">
-      <div className="space-y-2">
-        <label htmlFor="phoneNumber" className="text-sm font-medium">
-          Phone Number
-        </label>
-        <input
-          id="phoneNumber"
-          type="tel"
-          value={phoneNumber}
-          onChange={(e) => setPhoneNumber(e.target.value)}
-          className="w-full p-2 border rounded-md"
-          required
-        />
-      </div>
-      <div className="space-y-2">
-        <label htmlFor="password" className="text-sm font-medium">
-          Password
-        </label>
-        <input
-          id="password"
-          type="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          className="w-full p-2 border rounded-md"
-          required
-        />
-      </div>
-      <button
-        type="submit"
-        className="w-full p-2 text-white bg-blue-600 rounded-md hover:bg-blue-700"
-      >
-        Login
-      </button>
-    </form>
-  );
-}
