+++ src/components/login/LoggedInRedirect.tsx
@@ -1,0 +1,76 @@
+import { motion } from "framer-motion";
+import { Card, CardContent } from "@/components/ui/card";
+import { CheckCircle2 } from "lucide-react";
+
+interface LoggedInRedirectProps {
+  message?: string;
+}
+
+export function LoggedInRedirect({
+  message = "Redirecting to dashboard...",
+}: LoggedInRedirectProps) {
+  return (
+    <div className="min-h-screen flex items-center justify-center relative overflow-hidden">
+      <div className="absolute inset-0 bg-gradient-to-br from-blue-50 via-indigo-50 to-purple-50">
+        <div className="absolute inset-0 bg-grid-slate-200 [mask-image:linear-gradient(0deg,white,rgba(255,255,255,0.6))]" />
+        <div className="absolute inset-0 bg-gradient-to-br from-transparent via-purple-100/30 to-blue-100/30 animate-gradient" />
+      </div>
+
+      <motion.div
+        initial={{ scale: 0.8, opacity: 0 }}
+        animate={{ scale: 1, opacity: 1 }}
+        transition={{ duration: 0.5, ease: "easeOut" }}
+        className="relative z-10"
+      >
+        <Card className="w-[350px] shadow-xl bg-white/80 backdrop-blur-sm">
+          <CardContent className="pt-6">
+            <div className="flex flex-col items-center space-y-4">
+              <motion.div
+                initial={{ scale: 0 }}
+                animate={{ scale: 1 }}
+                transition={{ delay: 0.2, type: "spring", stiffness: 200 }}
+              >
+                <div className="rounded-full bg-green-100 p-3">
+                  <CheckCircle2 className="w-8 h-8 text-green-600" />
+                </div>
+              </motion.div>
+
+              <motion.div
+                initial={{ y: 20, opacity: 0 }}
+                animate={{ y: 0, opacity: 1 }}
+                transition={{ delay: 0.3 }}
+                className="text-center space-y-2"
+              >
+                <h2 className="text-2xl font-semibold text-gray-800">
+                  Welcome Back!
+                </h2>
+                <p className="text-gray-600">You&apos;re already logged in</p>
+              </motion.div>
+
+              <motion.div
+                initial={{ opacity: 0 }}
+                animate={{ opacity: 1 }}
+                transition={{ delay: 0.5 }}
+                className="w-full"
+              >
+                <div className="relative pt-4">
+                  <div className="w-full bg-gray-200 rounded-full h-1.5">
+                    <motion.div
+                      initial={{ width: "0%" }}
+                      animate={{ width: "100%" }}
+                      transition={{ duration: 2, ease: "easeInOut" }}
+                      className="bg-green-500 h-1.5 rounded-full"
+                    />
+                  </div>
+                </div>
+                <p className="text-sm text-gray-500 text-center mt-2">
+                  {message}
+                </p>
+              </motion.div>
+            </div>
+          </CardContent>
+        </Card>
+      </motion.div>
+    </div>
+  );
+}
