--- src/providers/auth-provider.tsx
@@ -1,83 +1,0 @@
-"use client";
-import { createContext, useContext, useEffect, useState } from "react";
-
-interface AuthContextType {
-  isAuthenticated: boolean;
-  user: { username: string } | null;
-  loading: boolean;
-  login: (phoneNumber: string, password: string) => Promise<boolean>;
-  logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | null>(null);
-
-export function AuthProvider({ children }: { children: React.ReactNode }) {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [user, setUser] = useState<{ username: string } | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    // Check local storage for access token on mount
-    const checkAuth = () => {
-      const accessToken = localStorage.getItem("access");
-      if (accessToken) {
-        // Assuming user data is included in token storage
-        setIsAuthenticated(true);
-        setUser({ username: "Admin" }); // Replace with actual user data if available
-      }
-      setLoading(false);
-    };
-
-    checkAuth();
-  }, []);
-
-  const login = async (phoneNumber: string, password: string) => {
-    try {
-      const response = await fetch(
-        `${process.env.NEXT_PUBLIC_API_URL}api/account/login/`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ phone_number: phoneNumber, password }),
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error("Login failed");
-      }
-
-      const data = await response.json();
-      localStorage.setItem("access", data.access); // Store access token
-      setIsAuthenticated(true);
-      setUser({ username: "Admin" }); // Replace with actual user data
-      return true;
-    } catch (error) {
-      console.error("Login error:", error);
-      return false;
-    }
-  };
-
-  const logout = () => {
-    localStorage.removeItem("access");
-    setUser(null);
-    setIsAuthenticated(false);
-  };
-
-  return (
-    <AuthContext.Provider
-      value={{ isAuthenticated, user, loading, login, logout }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-}
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
